Investa-Backend/requirements.txt
# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0

# Database
PyMongo==4.5.0
pymongo[srv]==4.5.0

# Data Validation
pydantic==2.4.2

# Authentication
firebase-admin==6.2.0

# HTTP Client
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# Production Server (optional)
gunicorn==21.2.0

# Development Tools (optional)
pytest==7.4.2
pytest-flask==1.2.0
